+++ src/main/kotlin/app/revanced/patcher/PatcherOptions.kt
@@ -1,10 +1,10 @@
 package app.revanced.patcher
 
 import app.revanced.patcher.data.ResourceContext
-import app.revanced.patcher.logging.Logger
 import app.revanced.patcher.logging.impl.NopLogger
 import brut.androlib.Config
 import java.io.File
+import java.util.logging.Logger
 
 /**
  * Options for ReVanced [Patcher].
@@ -12,15 +12,19 @@ import java.io.File
  * @param resourceCachePath The path to the directory to use for caching resources.
  * @param aaptBinaryPath The path to a custom aapt binary.
  * @param frameworkFileDirectory The path to the directory to cache the framework file in.
- * @param logger A [Logger].
+ * @param unusedLogger The logger to use for logging.
  */
-data class PatcherOptions(
+data class PatcherOptions
+@Deprecated("Use the constructor without the logger parameter instead")
+constructor(
     internal val inputFile: File,
     internal val resourceCachePath: File = File("revanced-resource-cache"),
     internal val aaptBinaryPath: String? = null,
     internal val frameworkFileDirectory: String? = null,
-    internal val logger: Logger = NopLogger
+    internal val unusedLogger: app.revanced.patcher.logging.Logger = NopLogger
 ) {
+    private val logger = Logger.getLogger(PatcherOptions::class.java.name)
+
     /**
      * The mode to use for resource decoding.
      * @see ResourceContext.ResourceDecodingMode
@@ -36,12 +40,32 @@ data class PatcherOptions(
         frameworkDirectory = frameworkFileDirectory
     }
 
+    /**
+     * Options for ReVanced [Patcher].
+     * @param inputFile The input file to patch.
+     * @param resourceCachePath The path to the directory to use for caching resources.
+     * @param aaptBinaryPath The path to a custom aapt binary.
+     * @param frameworkFileDirectory The path to the directory to cache the framework file in.
+     */
+    constructor(
+        inputFile: File,
+        resourceCachePath: File = File("revanced-resource-cache"),
+        aaptBinaryPath: String? = null,
+        frameworkFileDirectory: String? = null,
+    ) : this(
+        inputFile,
+        resourceCachePath,
+        aaptBinaryPath,
+        frameworkFileDirectory,
+        NopLogger
+    )
+
     fun recreateResourceCacheDirectory() = resourceCachePath.also {
         if (it.exists()) {
             logger.info("Deleting existing resource cache directory")
 
             if (!it.deleteRecursively())
-                logger.error("Failed to delete existing resource cache directory")
+                logger.severe("Failed to delete existing resource cache directory")
         }
 
         it.mkdirs()