+++ src/main/kotlin/app/revanced/patcher/logging/Logger.kt
@@ -1,5 +1,6 @@
 package app.revanced.patcher.logging
 
+@Deprecated("This will be removed in a future release")
 interface Logger {
     fun error(msg: String) {}
     fun warn(msg: String) {}